Refactor AdminUserPage to improve type imports and state management
- Updated import statements in AdminUserPage and its test file to source User and related types from '../types/users' for better organization.
- Enhanced state management in AdminUserPage by explicitly typing the form data state as CreateUserDTO or UpdateUserDTO, improving type safety and clarity.
